requirements.txt
Django>=5.1,<5.2
wagtail>=6.4,<6.5
psycopg2-binary>=2.9.10
boto3==1.34.97
django-storages==1.14.0
django-autoslug==1.9.9